Adding ottoman utils test
